feat(push notifications): change is_recurring boolean to recurring integer
- with having recurring as integer we can save 1 to x representing
  "each week" (1) or "every month" (4)
- fixed `notification_pushable_id` to bigint as it is an ID

SVA-710

+++ db/migrate/20221125161012_create_notification_pushes.rb
@@ -4,7 +4,7 @@ class CreateNotificationPushes < ActiveRecord::Migration[5.2]
   def change
     create_table :notification_pushes do |t|
       t.string :notification_pushable_type
-      t.integer :notification_pushable_id
+      t.bigint :notification_pushable_id
       t.datetime :once_at
       t.boolean :is_recurring, default: false
       t.time :monday_at
